+++ tests/unit/async_/work/test_session.py
@@ -23,6 +23,7 @@ import pytest
 from neo4j import (
     AsyncSession,
     AsyncTransaction,
+    Bookmarks,
     SessionConfig,
     unit_of_work,
 )
@@ -170,20 +171,25 @@ async def test_closes_connection_after_tx_commit(pool, test_run_args):
 
 
 @pytest.mark.parametrize(
-    "bookmarks",
+    "bookmark_values",
     (None, [], ["abc"], ["foo", "bar"], {"a", "b"}, ("1", "two"))
 )
 @mark_async_test
-async def test_session_returns_bookmark_directly(pool, bookmarks):
+async def test_session_returns_bookmarks_directly(pool, bookmark_values):
+    if bookmark_values is not None:
+        bookmarks = Bookmarks.from_raw_values(bookmark_values)
+    else:
+        bookmarks = Bookmarks()
     async with AsyncSession(
         pool, SessionConfig(bookmarks=bookmarks)
     ) as session:
-        if bookmarks is None:
-            assert await session.last_bookmarks() == []
-        elif isinstance(bookmarks, set):
-            assert sorted(await session.last_bookmarks()) == sorted(bookmarks)
+        ret_bookmarks = (await session.last_bookmarks())
+        assert isinstance(ret_bookmarks, Bookmarks)
+        ret_bookmarks = ret_bookmarks.raw_values
+        if bookmark_values is None:
+            assert ret_bookmarks == frozenset()
         else:
-            assert await session.last_bookmarks() == list(bookmarks)
+            assert ret_bookmarks == frozenset(bookmark_values)
 
 
 @pytest.mark.parametrize(("query", "error_type"), (